+++ ring/alembic/env.py
@@ -7,6 +7,9 @@ from ring.config import get_config
 from ring.letters.models.default_question import (
     DefaultQuestion,  # type: ignore # noqa: F401
 )
+from ring.letters.models.default_question_model import (
+    DefaultQuestion,  # type: ignore # noqa: F401
+)
 from ring.letters.models.letter_model import (
     Letter,  # type: ignore # noqa: F401
 )
@@ -18,7 +21,7 @@ from ring.letters.models.response_model import (
 )
 from ring.parties.models.group_model import Group  # type: ignore # noqa: F401
 from ring.parties.models.invite_model import (
-    Invite,  # type: ignore # noqa: F401
+    Invite,  # type: ignore # noqa: F401  # type: ignore # noqa: F401
 )
 from ring.parties.models.one_time_token_model import (
     OneTimeToken,  # type: ignore # noqa: F401
